requirements.txt
asgiref==3.8.1
Django==5.1.4
django-appconf==1.1.0
django-ckeditor-5==0.2.15
django-cors-headers==4.6.0
django-filter==24.3
django-jazzmin==3.0.1
django-parler==2.3
django-select2==8.1.2
django-tinymce==4.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.8
gunicorn==23.0.0
inflection==0.5.1
modeltranslation==0.25
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
pytz==2024.2
PyYAML==6.0.2
requests==2.31.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
python-dotenv==1.1.1
langchain-core==0.3.69
langchain_groq==0.3.6
elevenlabs==2.7.1
pydub==0.25.1